pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inquiry-lead-hunter"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client",
    "google-auth",
    "anthropic",
    "requests",
    "python-dotenv",
    "pyyaml",
    "tenacity",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]
cloud = [
    "functions-framework",
]

[project.scripts]
inquiry-lead-hunter = "inquiry_lead_hunter.main:main"

[tool.setuptools.packages.find]
where = ["src"]
